refactor: move custom codeactions to its folder
