+++ typescript/src/codeActions/getCodeActions.ts
@@ -1,8 +1,8 @@
 import { compact } from '@zardoy/utils'
 import { findChildContainingPosition } from '../utils'
-import objectSwapKeysAndValues from './objectSwapKeysAndValues'
-import changeStringReplaceToRegex from './changeStringReplaceToRegex'
-import toggleBraces from './toggleBraces'
+import objectSwapKeysAndValues from './custom/objectSwapKeysAndValues'
+import changeStringReplaceToRegex from './custom/changeStringReplaceToRegex'
+import toggleBraces from './custom/toggleBraces'
 
 type SimplifiedRefactorInfo =
     | {
